+++ components/contribution/RNBIDHeader.tsx
@@ -1,0 +1,12 @@
+type Props = {
+  rnbId: string;
+};
+
+export default function RNBIDheader({ rnbId }: Props) {
+  return (
+    <>
+      <span>Identifiant RNB</span>
+      <span>{rnbId}</span>
+    </>
+  );
+}
