+++ styles/contribution/editPanel.module.scss
@@ -1,4 +1,4 @@
-@import "vars.scss";
+@import "../vars.scss";
 
 $input-height: 52px;
 $spacing: 12px;
@@ -15,11 +15,12 @@ $shell-top: $actions-top + $spacing + $actions-height;
     z-index: 100;
     width: 400px;
 
-    background-color: red;
+    // TODO: Adapt once we have buttons
+    background-color: transparent;
+    pointer-events: none;
 }
 
 .shell {
-    // The .shell is the invisible container containing actions and panel
     position:absolute;
 
     top: $shell-top;
@@ -31,12 +32,10 @@ $shell-top: $actions-top + $spacing + $actions-height;
 }
 
 .content {
-    // Actual panel content
     background:white;
     box-shadow:0 0 0 2px rgba(0,0,0,.1);
     
-    border-top-left-radius: 4px;
-    border-top-right-radius: 4px;
+    border-radius: 4px;
     display: flex;
     flex-direction: column;
     max-height: 100%;
